pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "zadavalnik"
version = "0.0.1"
description = "Questioning bot"
authors = [
    { name = "Anton Lobachev", email = "anvlobachev@gmail.com" }
]
# license = { file = "LICENSE" }
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-telegram-bot",
    "numpy",
    "requests",
    "python-dotenv",
]

[tool.setuptools]
packages = {find = {where = ["src"]}}


[project.urls]
# Homepage = "https://github.com/username/my_package"
# Documentation = "https://github.com/username/my_package#readme"
# Issues = "https://github.com/username/my_package/issues"

# [tool.setuptools_scm]
# write_to = "my_package/_version.py"